requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.2.0
networkx==2.4
numpy==1.18.1
pandas==1.0.1
Pillow==7.1.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sklearn==0.0
